+++ R/conduct_parent_to_helpers.R
@@ -24,22 +24,13 @@
 #'
 #' @family helper functions
 #' @noRd
-#' @examplesIf "ETN" %in% odbc::odbcListDataSources()$name
-#' # When API is FALSE, this package forwards function calls directly to
-#' # etnservice
-#'
-#' # These two calls are identical. They run the same code (locally). Both will
-#' # only work when there is a local connection to the etn database.
-#' etnservice::list_acoustic_tag_ids()
-#' list_acoustic_tag_ids(api = FALSE)
-#'
-conduct_parent_to_helpers <- function(api,
+conduct_parent_to_helpers <- function(protocol = c("opencpu", "localdb"),
                                       ignored_arguments = NULL,
                                       ...) {
   # Check arguments
-  assertthat::assert_that(assertthat::is.flag(api))
+  protocol <- rlang::arg_match(protocol)
   assertthat::assert_that(is.character(ignored_arguments) |
-                            is.null(ignored_arguments))
+    is.null(ignored_arguments))
 
   # Lock in the name of the parent function
   function_identity <-
@@ -56,41 +47,44 @@ conduct_parent_to_helpers <- function(api,
       )
     ]
 
-  if (api) {
-    # Forward arguments to API via helper.
-    out <- do.call(
-      forward_to_api,
-      list(function_identity = function_identity,
-           payload = arguments_to_pass,
-           ...)
-    )
-  } else {
-    # Check if the local version of etnservice matches the one deployed on
-    # OpenCPU.
-    if (!etnservice_version_matches()) {
-      deployed_version <- get_etnservice_version(api = TRUE)
+  switch(protocol,
+    opencpu = {
+      # Forward arguments to API via helper.
+      do.call(
+        forward_to_api,
+        list(
+          function_identity = function_identity,
+          payload = arguments_to_pass,
+          ...
+        )
+      )
+    },
+    localdb = {
+      # Check if the local version of etnservice matches the one deployed on
+      # OpenCPU.
+      if (!etnservice_version_matches()) {
+        deployed_version <- get_etnservice_version(which = "opencpu")
 
-      rlang::check_installed(
-        "etnservice",
-        version = deployed_version,
-        # Ensure the exact version is installed, and not an even more recent
-        # version (In case OpenCPU is lagging on the Github released version).
-        compare = "==",
-        reason =
-          paste(
-            "\nThere is a newer version of etnservice available",
-            "please update",
-            "to avoid differences between local and API queries."
-          )
+        rlang::check_installed(
+          "etnservice",
+          version = deployed_version,
+          # Ensure the exact version is installed, and not an even more recent
+          # version (In case OpenCPU is lagging on the Github released version).
+          compare = "==",
+          reason =
+            paste(
+              "\nThere is a newer version of etnservice available",
+              "please update",
+              "to avoid differences between local and API queries."
+            )
+        )
+      }
+      do.call(utils::getFromNamespace(function_identity, ns = "etnservice"),
+        args = append(arguments_to_pass,
+          list(credentials = get_credentials()),
+          after = 0
+        )
       )
     }
-
-    out <- do.call(utils::getFromNamespace(function_identity, ns = "etnservice"),
-                   args = append(arguments_to_pass,
-                                 list(credentials = get_credentials()),
-                                 after = 0)
-    )
-  }
-
-  return(out)
+  )
 }