+++ R/get_acoustic_projects.R
@@ -19,14 +19,13 @@
 #' # Get a specific acoustic project
 #' get_acoustic_projects(acoustic_project_code = "demer")
 get_acoustic_projects <- function(connection,
-                                  acoustic_project_code = NULL,
-                                  api = TRUE){
+                                  acoustic_project_code = NULL) {
   # Check arguments
   # The connection argument has been depreciated
   if (lifecycle::is_present(connection)) {
     deprecate_warn_connection()
   }
   # Either use the API, or the SQL helper.
-  out <- conduct_parent_to_helpers(api)
+  out <- conduct_parent_to_helpers(protocol = select_protocol())
   return(out)
 }