+++ R/list_scientific_names.R
@@ -5,14 +5,12 @@
 #'   `common.animal_release`.
 #'
 #' @export
-list_scientific_names <- function(connection,
-                                  api = TRUE) {
+list_scientific_names <- function(connection) {
   # Check arguments
   # The connection argument has been depreciated
   if (lifecycle::is_present(connection)) {
     deprecate_warn_connection()
   }
   # Either use the API, or the SQL helper.
-  out <- conduct_parent_to_helpers(api, json = TRUE)
-  return(out)
+  conduct_parent_to_helpers(protocol = select_protocol(), json = TRUE)
 }