+++ tests/testthat/test-get_animal_projects.R
@@ -4,7 +4,7 @@ test_that("get_animal_projects() returns a tibble", {
   df <- get_animal_projects()
   expect_s3_class(df, "data.frame")
   expect_s3_class(df, "tbl")
-  df_sql <- get_animal_projects(api = FALSE)
+  df_sql <- get_animal_projects()
   expect_s3_class(df_sql, "data.frame")
   expect_s3_class(df_sql, "tbl")
 })