+++ tests/testthat/test-list_deployment_ids.R
@@ -1,5 +1,7 @@
 test_that("list_deployment_ids() returns unique list of values using api", {
-  vcr::use_cassette("list_deployment_ids", {vector <- list_deployment_ids()})
+  vcr::use_cassette("list_deployment_ids", {
+    vector <- list_deployment_ids()
+  })
 
   expect_type(vector, "character")
   expect_false(any(duplicated(vector)))
@@ -10,7 +12,7 @@ test_that("list_deployment_ids() returns unique list of values using api", {
 
 test_that("list_deployment_ids() returns unique list of values using local db", {
   skip_if_not_localdb()
-  vector_sql <- list_deployment_ids(api = FALSE)
+  vector_sql <- list_deployment_ids()
 
   expect_type(vector_sql, "character")
   expect_false(any(duplicated(vector_sql)))