+++ tests/testthat/test-write_dwc.R
@@ -5,7 +5,7 @@ test_that("[SQL] write_dwc() can write csv files to a path", {
   unlink(out_dir, recursive = TRUE)
   dir.create(out_dir)
   suppressMessages(
-    write_dwc(animal_project_code = "2014_demer", directory = out_dir, api = FALSE)
+    write_dwc(animal_project_code = "2014_demer", directory = out_dir)
   )
 
   expect_identical(
@@ -18,7 +18,7 @@ test_that("[SQL] write_dwc() can return data as list of tibbles rather than file
   skip_if_not_localdb()
 
   result <- suppressMessages(
-    write_dwc(animal_project_code = "2014_demer", directory = NULL, api = FALSE)
+    write_dwc(animal_project_code = "2014_demer", directory = NULL)
   )
 
   expect_identical(names(result), "dwc_occurrence")
@@ -29,7 +29,7 @@ test_that("[SQL] write_dwc() returns the expected Darwin Core terms as columns",
   skip_if_not_localdb()
 
   result <- suppressMessages(
-    write_dwc(animal_project_code = "2014_demer", directory = NULL, api = FALSE)
+    write_dwc(animal_project_code = "2014_demer", directory = NULL)
   )
 
   expect_identical(
@@ -68,71 +68,75 @@ test_that("[SQL] write_dwc() returns the expected Darwin Core terms as columns",
   )
 })
 
-  # cached version of API response for 2014_demer
-  vcr::use_cassette("2014_demer_dwc", {
-    # Request without writing to disk
-    demer_dwc <- write_dwc(
-      animal_project_code = "2014_demer",
-      directory = NULL,
-      api = TRUE
-    )
-  })
-
-  test_that("[API] write_dwc() can write csv files to a path", {
-    out_dir <- file.path(tempdir(), "dwc")
-    unlink(out_dir, recursive = TRUE)
-    dir.create(out_dir)
-    # Request that writes to disk
-    suppressMessages(write_dwc(
-      animal_project_code = "2014_demer",
-      directory = out_dir,
-      api = TRUE
-    ))
-
-    expect_identical(list.files(out_dir, pattern = "*.csv"), "dwc_occurrence.csv")
-  })
-
-  test_that("[API] write_dwc() can return data as list of tibbles rather than files", {
-    expect_identical(names(demer_dwc), "dwc_occurrence")
-    expect_s3_class(demer_dwc$dwc_occurrence, "tbl")
-  })
-
-  test_that("[API] write_dwc() returns the expected Darwin Core terms as columns", {
-    expect_identical(
-      colnames(demer_dwc$dwc_occurrence),
-      c(
-        "type",
-        "license",
-        "rightsHolder",
-        "datasetID",
-        "institutionCode",
-        "collectionCode",
-        "datasetName",
-        "basisOfRecord",
-        "dataGeneralizations",
-        "occurrenceID",
-        "sex",
-        "lifeStage",
-        "occurrenceStatus",
-        "organismID",
-        "organismName",
-        "eventID",
-        "parentEventID",
-        "eventDate",
-        "samplingProtocol",
-        "eventRemarks",
-        "locationID",
-        "locality",
-        "decimalLatitude",
-        "decimalLongitude",
-        "geodeticDatum",
-        "coordinateUncertaintyInMeters",
-        "scientificNameID",
-        "scientificName",
-        "kingdom"
+# Force protocol to be openCPU, cache a API response for 2014_demer
+withr::with_envvar(
+  c("ETN_PROTOCOL" = "opencpu"),
+  {
+    vcr::use_cassette("2014_demer_dwc", {
+      # Request without writing to disk
+      demer_dwc <- write_dwc(
+        animal_project_code = "2014_demer",
+        directory = NULL
       )
+    })
+  }
+)
+
+test_that("[API] write_dwc() can write csv files to a path", {
+  # Force using the OpenCPU API
+  withr::local_envvar("ETN_PROTOCOL" = "opencpu")
+  out_dir <- withr::local_tempdir()
+  # Request that writes to disk
+  suppressMessages(write_dwc(
+    animal_project_code = "2014_demer",
+    directory = out_dir
+  ))
+
+  expect_identical(list.files(out_dir, pattern = "*.csv"), "dwc_occurrence.csv")
+})
+
+test_that("[API] write_dwc() can return data as list of tibbles rather than files", {
+  expect_identical(names(demer_dwc), "dwc_occurrence")
+  expect_s3_class(demer_dwc$dwc_occurrence, "tbl")
+})
+
+test_that("[API] write_dwc() returns the expected Darwin Core terms as columns", {
+  expect_identical(
+    colnames(demer_dwc$dwc_occurrence),
+    c(
+      "type",
+      "license",
+      "rightsHolder",
+      "datasetID",
+      "institutionCode",
+      "collectionCode",
+      "datasetName",
+      "basisOfRecord",
+      "dataGeneralizations",
+      "occurrenceID",
+      "sex",
+      "lifeStage",
+      "occurrenceStatus",
+      "organismID",
+      "organismName",
+      "eventID",
+      "parentEventID",
+      "eventDate",
+      "samplingProtocol",
+      "eventRemarks",
+      "locationID",
+      "locality",
+      "decimalLatitude",
+      "decimalLongitude",
+      "geodeticDatum",
+      "coordinateUncertaintyInMeters",
+      "scientificNameID",
+      "scientificName",
+      "kingdom"
     )
-  })
+  )
+})
+
 test_that("[SQL] write_dwc() supports uppercase animal_project_codes", {
   skip_if_not_localdb()
   result <- suppressMessages(
@@ -148,9 +152,11 @@ test_that("[SQL] write_dwc() supports uppercase animal_project_codes", {
 test_that("[SQL] write_dwc() allows setting of rights_holder", {
   skip_if_not_localdb()
   result <- suppressMessages(
-    write_dwc(animal_project_code = "2014_demer",
-              rights_holder = "my_rightholder",
-              directory = NULL)
+    write_dwc(
+      animal_project_code = "2014_demer",
+      rights_holder = "my_rightholder",
+      directory = NULL
+    )
   )
 
   expect_identical(
@@ -162,8 +168,10 @@ test_that("[SQL] write_dwc() allows setting of rights_holder", {
 test_that("[SQL] write_dwc() returns an empty column for rights holder by default", {
   skip_if_not_localdb()
   result <- suppressMessages(
-    write_dwc(animal_project_code = "2014_demer",
-              directory = NULL)
+    write_dwc(
+      animal_project_code = "2014_demer",
+      directory = NULL
+    )
   )
 
   expect_identical(